ppErrorPreprocess Make use of the existing Error::Preprocess
Before this commit, inputs with invalid pp syntax are reported using
`Error::Parse` which is also used to report invalid post-pp syntax.
In svlint/svls, `Error::Parse` is reported to the user as "parse error"
without specifying if that's pre or post pp.
This commit makes it possible for future versions of svlint/svls to report
"preprocess error", thus helping users to identify pp issues.

- `Error::Preprocess` appears to be defined but unused.
- Add the argument `Option<(PathBuf, usize)>` to `Error::Preprocess`, same
  as `Error::Parse`.
- Update `preprocess_str()` to use `Error::Preprocess` in all 5 instances.